+++ src/JlWrap/JlWrap.jl
@@ -20,6 +20,7 @@ include("base.jl")
 include("any.jl")
 include("io.jl")
 include("objectarray.jl")
+include("collection.jl")
 include("array.jl")
 include("vector.jl")
 include("dict.jl")
@@ -27,10 +28,11 @@ include("set.jl")
 include("callback.jl")
 
 function __init__()
-    Cjl.C.with_gil() do 
+    Cjl.C.with_gil() do
         init_base()
         init_any()
         init_io()
+        init_collection()
         init_array()
         init_vector()
         init_dict()
