+++ src/JlWrap/any.jl
@@ -355,21 +355,6 @@ function init_any()
             return self._jl_callmethod($(pyjl_methodnum(pyjlany_hash)))
         def __next__(self):
             return self._jl_callmethod($(pyjl_methodnum(pyjliter_next)))
-    class _JlContainerMixin(_JlReprMixin):
-        __slots__ = ()
-        def __len__(self):
-            return self._jl_callmethod($(pyjl_methodnum(pyint ∘ length)))
-        def __bool__(self):
-            return self._jl_callmethod($(pyjl_methodnum(pybool ∘ !isempty)))
-        def __iter__(self):
-            return self._jl_callmethod($(pyjl_methodnum(pyjliter ∘ Iterator)))
-        def __hash__(self):
-            return self._jl_callmethod($(pyjl_methodnum(pyjlany_hash)))
-        def __eq__(self, other):
-            if isinstance(self, type(other)) or isinstance(other, type(self)):
-                return self._jl_callmethod($(pyjl_methodnum(pyjlmixin_eq_bool)), other)
-            else:
-                return NotImplemented
     class Jl(JlBase, _JlReprMixin):
         __slots__ = ()
         def __str__(self):
